Notifi: Support hyperlink in history view/ refactor history view
