+++ packages/web/package.json
@@ -45,8 +45,9 @@
     "@keplr-wallet/wc-client": "0.10.24-ibc.go.v7.hot.fix",
     "@metamask/onboarding": "^1.0.1",
     "@next/bundle-analyzer": "^12.1.6",
-    "@notifi-network/notifi-frontend-client": "0.84.0",
-    "@notifi-network/notifi-react-card": "0.84.0",
+    "@notifi-network/notifi-frontend-client": "0.87.0",
+    "@notifi-network/notifi-react-card": "0.87.0",
+    "dompurify": "^3.0.6",
     "@osmosis-labs/keplr-hooks": "0.10.24-ibc.go.v7.hot.fix",
     "@osmosis-labs/keplr-stores": "0.10.24-ibc.go.v7.hot.fix",
     "@osmosis-labs/math": "^5.1.0",
@@ -156,6 +157,7 @@
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
-    "use-debugger-hooks": "^1.3.0"
+    "use-debugger-hooks": "^1.3.0",
+    "@types/dompurify": "^3.0.4"
   }
 }
